chore: Code refactor
