+++ src/controllers/authController.js
@@ -48,7 +48,7 @@ exports.login = asyncHandler(async (req, res, next) => {
     });
   } catch (err) {
     return res.status(500).json({
-      error: err.error,
+      error: err,
     });
   }
 });
